feat: 상품 상세 기능 추가 구현

+++ src/main/java/cholog/wiseshop/api/product/service/ProductService.java
@@ -1,9 +1,11 @@
 package cholog.wiseshop.api.product.service;
 
 import cholog.wiseshop.api.campaign.dto.request.CreateCampaignRequest.CreateProductRequest;
-import cholog.wiseshop.api.product.dto.request.ModifyProductPriceRequest;
+import cholog.wiseshop.api.product.dto.request.ModifyProductPriceAndStockRequest;
 import cholog.wiseshop.api.product.dto.request.ModifyProductRequest;
 import cholog.wiseshop.api.product.dto.response.ProductResponse;
+import cholog.wiseshop.db.campaign.Campaign;
+import cholog.wiseshop.db.campaign.CampaignRepository;
 import cholog.wiseshop.db.product.Product;
 import cholog.wiseshop.db.product.ProductRepository;
 import cholog.wiseshop.db.stock.Stock;
@@ -17,22 +19,11 @@ import org.springframework.transaction.annotation.Transactional;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final CampaignRepository campaignRepository;
 
-    public ProductService(ProductRepository productRepository) {
+    public ProductService(ProductRepository productRepository, CampaignRepository campaignRepository) {
         this.productRepository = productRepository;
-    }
-
-    // TODO: 안쓰면 제거
-    public Long createProduct(CreateProductRequest request) {
-        Stock stock = new Stock(request.totalQuantity());
-        Product product = Product.builder()
-            .name(request.name())
-            .description(request.description())
-            .price(request.price())
-            .stock(stock)
-            .build();
-        Product createdProduct = productRepository.save(product);
-        return createdProduct.getId();
+        this.campaignRepository = campaignRepository;
     }
 
     @Transactional(readOnly = true)
@@ -49,17 +40,23 @@ public class ProductService {
         productRepository.save(existedProduct);
     }
 
-    public void modifyProductPrice(Long productId, ModifyProductPriceRequest request) {
+    public void modifyProductPriceAndStock(Long productId, ModifyProductPriceAndStockRequest request) {
         Product existedProduct = productRepository.findById(productId)
-            .orElseThrow(
-                () -> new WiseShopException(WiseShopErrorCode.MODIFY_PRICE_PRODUCT_NOT_FOUND));
-        existedProduct.modifyPrice(request.price());
+            .orElseThrow(() -> new WiseShopException(WiseShopErrorCode.MODIFY_PRICE_PRODUCT_NOT_FOUND));
+        existedProduct.modifyPriceAndStock(request.price(), request.totalQuantity());
         productRepository.save(existedProduct);
     }
 
     public void deleteProduct(Long id) {
-        productRepository.findById(id)
+        Product product = productRepository.findById(id)
             .orElseThrow(() -> new WiseShopException(WiseShopErrorCode.PRODUCT_NOT_FOUND));
+        if (product.getCampaign().isNotWaiting()) {
+            throw new WiseShopException(WiseShopErrorCode.INVALID_CAMPAIGN_DELETE_STATE);
+        }
+        Campaign campaign = product.getCampaign();
         productRepository.deleteById(id);
+        if (productRepository.findAllByCampaign(campaign).isEmpty()) {
+            campaignRepository.delete(campaign);
+        }
     }
 }
